ADD the new requirements and some updates

+++ src/main/java/com/example/fawranwebservice/Discounts/DiscountCTR.java
@@ -25,19 +25,26 @@ public class DiscountCTR {
 
     @PostMapping("/createOverAll")
     public Response createOverAllDiscount(@RequestBody Discount discount){
-        discountService.addOverAllDiscount(discount);
-        return new Response(true, "Discount ADDED");
+        if(discountService.checkAdmin()) {
+            discountService.addOverAllDiscount(discount);
+            return new Response(true, "Discount ADDED");
+        }
+        return new Response(false, "YOU ARE NOT A ADMIN");
     }
     @PostMapping("/createSpecific")
     public Response createSpecificDiscount(@RequestBody Map<String, Discount> discount){
-        discountService.addSpecificDiscount(discount);
-        return new Response(true, "Discount ADDED",discount);
+        if(discountService.checkAdmin()) {
+            discountService.addSpecificDiscount(discount);
+            return new Response(true, "Discount ADDED", discount);
+        }
+        return new Response(false, "YOU ARE NOT A ADMIN");
     }
     @PutMapping("/delete")
     public Response deleteDiscount(@RequestBody Discount discountDescription){
-        discountService.deleteDiscount(discountDescription.getDescription());
-        return null;
+        if(discountService.checkAdmin()) {
+            discountService.deleteDiscount(discountDescription.getDescription());
+            return new Response(true, "Discount Removed");
+        }
+        return new Response(false, "YOU ARE NOT A ADMIN");
     }
-
-
 }
